+++ OrchestratorV1/Tests/task.getParameters.ts
@@ -9,14 +9,16 @@ const tmr: mr.TaskMockRunner = new mr.TaskMockRunner(path.join(__dirname, "mock.
 MockInput(tmr, [
 
     "ReleaseStrategy",
-    "StageStrategy",
-    "ArtifactStrategy",
     "TargetProject",
     "TargetDefinition",
+    "DefinitionStagesFilter",
     "TargetDefinitionStages",
     "TargetRelease",
     "TargetReleaseStages",
-    "TargetArtifactVersion",
+    "ArtifactTagFilter",
+    "ArtifactTagName",
+    "SourceBranchFilter",
+    "SourceBranchName"
 
 ]);